+++ src/chocolatey/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2022 Adam Karpierz
 # SPDX-License-Identifier: Zlib
 
-from .__about__ import * ; del __about__  # noqa
+from .__about__ import * ; del __about__  # type: ignore[name-defined]  # noqa
 
-from ._chocolatey     import * ; del _chocolatey  # noqa
-from ._chocolatey_cmd import * ; del _chocolatey_cmd  # noqa
+from ._chocolatey     import * ; del _chocolatey      # type: ignore[name-defined]  # noqa
+from ._chocolatey_cmd import * ; del _chocolatey_cmd  # type: ignore[name-defined]  # noqa